requirements.txt
datasets==2.4.0
evaluate==0.2.2
flask==2.2.2
huggingface-hub==0.9.1
numpy==1.23.1
numpy-base==1.23.1
pytorch==1.10.1
tokenizers==0.12.1
torchaudio==0.10.1
torchvision==0.11.2
tqdm==4.64.1
transformers==4.22.0.dev0

# git clone --single-branch --branch fast_tokenizers_BARTpho_PhoBERT_BERTweet https://github.com/datquocnguyen/transformers.git
# cd transformers
# pip3 install -e .